Build with typescript only.
