--- module.cjs
@@ -1,6 +1,0 @@
-// CommonJS proxy to bypass jiti transforms from nuxt 2 and using native ESM
-module.exports = function(...args) {
-  return import('./dist/module.mjs').then(m => m.default.call(this, ...args))
-}
-
-module.exports.meta = require('./package.json')
