+++ src/types/index.ts
@@ -2,7 +2,7 @@
 import type Vue from 'vue'
 import {} from '@nuxt/types'
 import { GraphqlMiddlewareConfig } from '../module'
-import { GraphqlMiddlewarePlugin } from '../runtime/middlewarePlugin'
+import { GraphqlMiddlewarePlugin } from '../templates/plugin'
 
 // @ts-ignore
 declare module 'vue/types/vue' {
