+++ tsconfig-cjs.json
@@ -1,0 +1,10 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "CommonJS",
+    "declaration": false,
+    "declarationDir": null,
+    "declarationMap": false,
+    "outDir": "./dist/cjs"
+  }
+}
