+++ tsconfig.json
@@ -3,27 +3,16 @@
     "target": "ESNext",
     "module": "ESNext",
     "moduleResolution": "node",
-    "lib": [
-      "DOM"
-    ],
+    "lib": ["DOM"],
     "esModuleInterop": true,
-    "resolveJsonModule": true,
     "strict": false,
-    "baseUrl": ".",
-    "outDir": "dist",
     "declaration": true,
     "declarationDir": "./dist/types",
     "declarationMap": true,
-    "types": [
-      "node",
-    ],
-    "paths": {
-      "~nuxtgraphqlmiddleware/*": ["src/runtime/*"],
-      "~nuxtgraphqlmiddleware": ["src/runtime"]
-    }
+    "removeComments": true,
+    "types": ["node"],
+    "outDir": "dist/esm"
   },
-  "include": [
-    "./src"
-  ],
-  "exclude": ["dist","node_modules"]
+  "include": ["./src"],
+  "exclude": ["dist", "node_modules"]
 }
